+++ src/main/java/ru/yandex/practicum/filmorate/model/Genre.java
@@ -5,10 +5,10 @@ import lombok.Data;
 @Data
 public class Genre {
 
-    private Byte id;
+    private int id;
     private String name;
 
-    public Genre(Byte id, String name) {
+    public Genre(int id, String name) {
         this.id = id;
         this.name = name;
     }
